Extend buildbundle to allow an optional meta value.
An optional value can be appended to the metadata
file that will be included in the signature. This
opaque value has no internal meaning to buildbundle
and only has context by follow-on tools and parsers.

Both buildsebundle and buildmmacbundle have leveraged
this extra value to allow policy to include the notion
of enforcing or permissive mode when shipped.

Change-Id: Ia66908ad98b2138818cfee79d3eaa20eb2597ebe
Signed-off-by: rpcraig <rpcraig@tycho.ncsc.mil>